First complete proofs of security (#10)
This PR finishes two proofs: one of completeness, and one of round-by-round knowledge soundness, for simplified versions of (a single round of) sumcheck.

TODOs: the proofs currently are pretty ad-hoc and brittle, so we need to build infrastructure around that. Also, the next step is to string these protocols together by describing & proving verified transformations (i.e. sequential composition, virtualization)

* current state of simp progress on sum-check

* type

* update in preparation for pulling from VCV's latest changes

* bump to v4.18.0-rc1

* got correctness proof working! (UniPoly still has errors)

* sorried out broken `UniPoly` proofs

* add more materials on rbr knowledge soundness of simple protocol

* finished proof of rbr knowledge soundness!

---------

Co-authored-by: Quang Dao <quangdao@umich.edu>